sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flatagents"
version = "0.4.0"
description = "A lightweight framework for building LLM-powered agents and composable state machines with pluggable backends."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { email = "memgrafter@gmail.com" }
]
keywords = ["llm", "agents", "ai", "framework", "litellm", "aisuite"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = ["pyyaml", "jinja2", "aiofiles", "httpx"]

[project.optional-dependencies]
litellm = ["litellm"]
aisuite = ["aisuite[all]"]
cel = ["cel-python"]
validation = ["jsonschema>=4.0"]
metrics = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
aws = ["boto3>=1.28.0"]
all = ["litellm", "aisuite[all]", "cel-python", "jsonschema>=4.0", "opentelemetry-api>=1.20.0", "opentelemetry-sdk>=1.20.0", "opentelemetry-exporter-otlp>=1.20.0", "boto3>=1.28.0"]

[project.urls]
Homepage = "https://github.com/memgrafter/flatagents"
Repository = "https://github.com/memgrafter/flatagents"
Issues = "https://github.com/memgrafter/flatagents/issues"

[tool.hatch.build.targets.wheel]
packages = ["flatagents"]
artifacts = [
    "flatagents/assets/*.d.ts",
    "flatagents/assets/*.json",
]
